Optimised calls to nest-coroutines.
Now, there is one level of indirection less per call, at the cost that only the root coroutine may be called. When entering a new coroutine, it is called directly (static), and when returning to the parent coroutine, the parent coroutine is called directly as well (not static though, with function pointer). Now, nest-coroutines are called through one pointer indirection and a function pointer.