Added CR_THROW and exception handling.
CR_THROW terminates the coroutine, but notifies the parent coroutine
that there was an error. CR_THROW must not be used in a coroutine that
has no parent coroutine. CR_CALL and CR_AWAIT now take an optional
argument containing error handling code. CR_CALL now takes the arguments
passed to the coroutine wrapped in parentheses:

    CR_CALL(coroutine, (arg, arg2), { puts("error occurred"); });
    CR_AWAIT(cv.wait(), { puts("error occurred"); });