Added release flag, cr_destroy(), AutoCoroutine.
By passing -DLIBCR_RELEASE=ON in CMake, release mode is enabled.
cr_destroy() is a function that has to be implemented in every
coroutine, and is called just before returning. util::AutoCoroutine<T>
is a wrapper coroutine that deletes itself and the wrapped coroutine
after execution finished. Although not optimal, this allows the dynamic
use of coroutines that do not delete themselves.